requirements.txt
requests==2.9.1
dj-database-url==0.4.0
Django==1.11
beautifulsoup4
gunicorn==19.4.5
psycopg2
whitenoise==2.0.6
django-filter==1.1